+++ core/src/com/scramble_like/game/game_object/Background.java
@@ -30,14 +30,14 @@ public class Background extends GameObject
     }
 
     @Override
-    public void Update(double DeltaTime)
+    public void Update(float DeltaTime)
     {
         if (!this.IsActive()) { return; }
 
-        this.getTransform().Translate((float) (DeltaTime * GameConstant.CAMERA_SPEED), 0);
+        this.getTransform().Translate(DeltaTime * GameConstant.CAMERA_SPEED, 0);
 
-        background_1.addOffset(new Vector2((float) (DeltaTime * GameConstant.BACKGROUND_SPEED), 0));
-        background_2.addOffset(new Vector2((float) (DeltaTime * GameConstant.BACKGROUND_SPEED), 0));
+        background_1.addOffset(new Vector2(DeltaTime * GameConstant.BACKGROUND_SPEED, 0));
+        background_2.addOffset(new Vector2(DeltaTime * GameConstant.BACKGROUND_SPEED, 0));
 
         if (camera.getPosition().x - background_1.getX() > length * 1.5f) { background_1.setX(background_2.getX() + length); }
         if (camera.getPosition().x - background_2.getX() > length * 1.5f) { background_2.setX(background_1.getX() + length); }
